+++ src/backend/utils/utils.hh
@@ -1,5 +1,5 @@
 #pragma once
 
-inline double sqr(double x) { return x * x; }
-inline double discr(double a, double b, double c) { return sqr(b) - (4 * a * c); }
-inline double abs_(double x) { return x < 0 ? -x : x; }
+inline float sqr(float x) { return x * x; }
+inline float discr(float a, float b, float c) { return sqr(b) - (4 * a * c); }
+inline float abs_(float x) { return x < 0 ? -x : x; }
