+++ src/backend/utils/vector3.hh
@@ -5,26 +5,26 @@
 
 struct Vector3
 {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 
     // Constructor & Destructor
     Vector3();
-    Vector3(double f);
-    Vector3(double x, double y, double z);
+    Vector3(float f);
+    Vector3(float x, float y, float z);
     Vector3(const Vector3& v);
 
     ~Vector3() = default;
 
     // Methods
-    double length_sqr();
-    double length();
+    float length_sqr();
+    float length();
 
-    double normalize();
+    float normalize();
     Vector3 norm() const;
 
-    double dot(const Vector3& v) const;
+    float dot(const Vector3& v) const;
     Vector3 cross(const Vector3& v) const;
 
     std::string to_string() const;
@@ -34,14 +34,14 @@ struct Vector3
     Vector3& operator+=(const Vector3& v);
     Vector3& operator-=(const Vector3& v);
     Vector3 operator-() const;
-    Vector3& operator*=(const double& f);
-    Vector3& operator/=(const double& f);
+    Vector3& operator*=(const float& f);
+    Vector3& operator/=(const float& f);
 };
 
 typedef Vector3 Point3;
 
 // Usefull operators on Vectors
-double dot(const Vector3& v1, const Vector3& v2);
+float dot(const Vector3& v1, const Vector3& v2);
 Vector3 cross(const Vector3& v1, const Vector3& v2);
 std::ostream& operator<<(std::ostream& out, const Vector3& vect);
 
@@ -67,14 +67,14 @@ inline Vector3 operator*(const Vector3& v1, const Vector3& v2)
                     v1.x * v2.y - v1.y * v2.x);
 }
 
-inline Vector3 operator*(const Vector3& v, double f)
+inline Vector3 operator*(const Vector3& v, float f)
 {
     return Vector3(v.x * f,
                    v.y * f,
                    v.z * f);
 }
 
-inline Vector3 operator*(double f, const Vector3& v)
+inline Vector3 operator*(float f, const Vector3& v)
 {
     return Vector3(f * v.x,
                    f * v.y,
@@ -88,14 +88,14 @@ inline Vector3 operator/(const Vector3& v1, const Vector3& v2)
                    v1.z / v2.z);
 }
 
-inline Vector3 operator/(const Vector3& v, double f)
+inline Vector3 operator/(const Vector3& v, float f)
 {
     return Vector3(v.x / f,
                    v.y / f,
                    v.z / f);
 }
 
-inline Vector3 operator/(double f, const Vector3& v)
+inline Vector3 operator/(float f, const Vector3& v)
 {
     return Vector3(f / v.x,
                    f / v.y,
@@ -104,7 +104,7 @@ inline Vector3 operator/(double f, const Vector3& v)
 
 inline bool operator==(const Vector3 &v1, const Vector3 &v2) 
 {
-    static const double delta = 0.001;
+    static const float delta = 0.001;
     return abs_(v1.x - v2.x) < delta &&
            abs_(v1.y - v2.y) < delta &&
            abs_(v1.z - v2.z) < delta;
